Worked out some adjustments I discussed with @isaacadams today.
