+++ src/views/components/ListView/index.js
@@ -1,0 +1,13 @@
+function ListView(props) {
+    return (
+        <div style={{ width: "90%" }}>
+            {props.itemList.map(({ id, name }) => (
+                <div key={id}>
+                    {name}
+                </div>
+            ))}
+        </div>
+    );
+}
+
+export default ListView;
