+++ src/views/components/LoadingScreen/index.js
@@ -1,0 +1,10 @@
+function LoadingScreen() {
+  return (
+    <div className="splash-screen">
+      Wait a moment while we load your app.
+      <div className="loading-dot">.</div>
+    </div>
+  );
+}
+
+export default LoadingScreen;
