--- src/views/pages/BackupWallet/BackupWallet.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function BackupWallet() {
-  return (
-    <div>
-      <h1>Backup Wallet</h1>
-    </div>
-  );
-};
-
-export default BackupWallet;
