+++ src/views/pages/BackupWallet/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function BackupWallet() {
+  return (
+    <div>
+      <h1>Backup Wallet</h1>
+    </div>
+  );
+};
+
+export default BackupWallet;
