--- src/views/pages/GenerateKeypair/GenerateKeypair.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function GenerateKeypair() {
-  return (
-    <div>
-      <h1>Generate RSA Keypair</h1>
-    </div>
-  );
-};
-
-export default GenerateKeypair;
