+++ src/views/pages/GenerateKeypair/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function GenerateKeypair() {
+  return (
+    <div>
+      <h1>Generate RSA Keypair</h1>
+    </div>
+  );
+};
+
+export default GenerateKeypair;
