--- src/views/pages/GenerateWallet/GenerateWallet.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function GenerateWallet() {
-  return (
-    <div>
-      <h1>Generate Wallet</h1>
-    </div>
-  );
-};
-
-export default GenerateWallet;
