+++ src/views/pages/GenerateWallet/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function GenerateWallet() {
+  return (
+    <div>
+      <h1>Generate Wallet</h1>
+    </div>
+  );
+};
+
+export default GenerateWallet;
