--- src/views/pages/ImportKeypair/ImportKeypair.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function ImportKeypair() {
-  return (
-    <div>
-      <h1>Import RSA Keypair</h1>
-    </div>
-  );
-};
-
-export default ImportKeypair;
