+++ src/views/pages/ImportKeypair/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function ImportKeypair() {
+  return (
+    <div>
+      <h1>Import RSA Keypair</h1>
+    </div>
+  );
+};
+
+export default ImportKeypair;
