--- src/views/pages/Navigation/Navigation.jsx
@@ -1,29 +1,0 @@
-import React from 'react';
-import './Navigation.styles.css';
-import { Link } from 'react-router-dom';
-
-function Navigation() {
-  return (
-    <nav
-      style={{
-        borderBottom: 'solid 1px',
-        paddingBottom: '1rem',
-      }}
-    >
-      <a>brand</a>
-      <ul>
-        <li>
-          <Link to="/">Home</Link>
-        </li>
-        <li>
-          <Link to="/profile">Profile</Link>
-        </li>
-        <li>
-          <Link to="/inbox">Inbox</Link>
-        </li>
-      </ul>
-    </nav>
-  );
-}
-
-export default Navigation;
