--- src/views/pages/Profile/Profile.jsx
@@ -1,5 +1,0 @@
-function Profile() {
-  return 'Profile';
-}
-
-export default Profile;
