+++ src/views/pages/Profile/index.js
@@ -1,0 +1,5 @@
+function Profile() {
+  return 'Profile';
+}
+
+export default Profile;
