--- src/views/pages/RestoreWallet/RestoreWallet.jsx
@@ -1,13 +1,0 @@
-import React  from "react";
-
-function RestoreWallet() {
-  return (
-    <div>
-      <h1>Restore Wallet</h1>
-      <textarea name="new_message" rows={5} cols={5} />
-      <button type="submit" />
-    </div>
-  );
-};
-
-export default RestoreWallet;
